pom.xml adjusted to build correct client zip, elements enums were moved to one package
