+++ excitebike/src/main/java/com/codenjoy/dojo/excitebike/client/ai/AISolver.java
@@ -25,8 +25,8 @@ package com.codenjoy.dojo.excitebike.client.ai;
 
 import com.codenjoy.dojo.client.Solver;
 import com.codenjoy.dojo.excitebike.client.Board;
-import com.codenjoy.dojo.excitebike.model.items.bike.Bike;
-import com.codenjoy.dojo.excitebike.model.items.bike.BikeType;
+import com.codenjoy.dojo.excitebike.model.items.Bike;
+import com.codenjoy.dojo.excitebike.model.elements.BikeType;
 import com.codenjoy.dojo.services.Dice;
 import com.codenjoy.dojo.services.Direction;
 import com.codenjoy.dojo.services.printer.CharElements;
@@ -38,17 +38,17 @@ import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.ACCELERATOR;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.FENCE;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.LINE_CHANGER_DOWN;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.LINE_CHANGER_UP;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.OBSTACLE;
-import static com.codenjoy.dojo.excitebike.model.items.bike.BikeType.BIKE_AT_LINE_CHANGER_DOWN;
-import static com.codenjoy.dojo.excitebike.model.items.bike.BikeType.BIKE_AT_LINE_CHANGER_UP;
-import static com.codenjoy.dojo.excitebike.model.items.bike.BikeType.OTHER_BIKE_AT_INHIBITOR;
-import static com.codenjoy.dojo.excitebike.model.items.bike.BikeType.OTHER_BIKE_AT_KILLED_BIKE;
-import static com.codenjoy.dojo.excitebike.model.items.bike.BikeType.OTHER_BIKE_AT_LINE_CHANGER_DOWN;
-import static com.codenjoy.dojo.excitebike.model.items.bike.BikeType.OTHER_BIKE_AT_LINE_CHANGER_UP;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.ACCELERATOR;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.FENCE;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.LINE_CHANGER_DOWN;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.LINE_CHANGER_UP;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.OBSTACLE;
+import static com.codenjoy.dojo.excitebike.model.elements.BikeType.BIKE_AT_LINE_CHANGER_DOWN;
+import static com.codenjoy.dojo.excitebike.model.elements.BikeType.BIKE_AT_LINE_CHANGER_UP;
+import static com.codenjoy.dojo.excitebike.model.elements.BikeType.OTHER_BIKE_AT_INHIBITOR;
+import static com.codenjoy.dojo.excitebike.model.elements.BikeType.OTHER_BIKE_AT_KILLED_BIKE;
+import static com.codenjoy.dojo.excitebike.model.elements.BikeType.OTHER_BIKE_AT_LINE_CHANGER_DOWN;
+import static com.codenjoy.dojo.excitebike.model.elements.BikeType.OTHER_BIKE_AT_LINE_CHANGER_UP;
 import static com.codenjoy.dojo.services.Direction.DOWN;
 import static com.codenjoy.dojo.services.Direction.RIGHT;
 import static com.codenjoy.dojo.services.Direction.UP;
