+++ excitebike/src/main/java/com/codenjoy/dojo/excitebike/model/GameField.java
@@ -23,7 +23,7 @@ package com.codenjoy.dojo.excitebike.model;
  */
 
 
-import com.codenjoy.dojo.excitebike.model.items.bike.Bike;
+import com.codenjoy.dojo.excitebike.model.items.Bike;
 
 import java.util.List;
 import java.util.Optional;
