+++ excitebike/src/main/java/com/codenjoy/dojo/excitebike/model/GameFieldImpl.java
@@ -25,8 +25,8 @@ package com.codenjoy.dojo.excitebike.model;
 
 import com.codenjoy.dojo.excitebike.model.items.Fence;
 import com.codenjoy.dojo.excitebike.model.items.Shiftable;
-import com.codenjoy.dojo.excitebike.model.items.bike.Bike;
-import com.codenjoy.dojo.excitebike.model.items.springboard.SpringboardElementType;
+import com.codenjoy.dojo.excitebike.model.items.Bike;
+import com.codenjoy.dojo.excitebike.model.elements.SpringboardElementType;
 import com.codenjoy.dojo.excitebike.services.Events;
 import com.codenjoy.dojo.excitebike.services.SettingsHandler;
 import com.codenjoy.dojo.excitebike.services.generation.GenerationOption;
@@ -51,13 +51,13 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.ACCELERATOR;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.INHIBITOR;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.LINE_CHANGER_DOWN;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.LINE_CHANGER_UP;
-import static com.codenjoy.dojo.excitebike.model.items.GameElementType.OBSTACLE;
-import static com.codenjoy.dojo.excitebike.model.items.bike.Bike.OTHER_BIKE_PREFIX;
-import static com.codenjoy.dojo.excitebike.model.items.bike.BikeType.BIKE_FALLEN;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.ACCELERATOR;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.INHIBITOR;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.LINE_CHANGER_DOWN;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.LINE_CHANGER_UP;
+import static com.codenjoy.dojo.excitebike.model.elements.GameElementType.OBSTACLE;
+import static com.codenjoy.dojo.excitebike.model.items.Bike.OTHER_BIKE_PREFIX;
+import static com.codenjoy.dojo.excitebike.model.elements.BikeType.BIKE_FALLEN;
 import static com.codenjoy.dojo.services.PointImpl.pt;
 import static java.util.Map.Entry.comparingByValue;
 import static java.util.stream.Collectors.groupingBy;
