+++ excitebike/src/main/java/com/codenjoy/dojo/excitebike/model/elements/BikeType.java
@@ -1,4 +1,4 @@
-package com.codenjoy.dojo.excitebike.model.items.bike;
+package com.codenjoy.dojo.excitebike.model.elements;
 
 /*-
  * #%L
