+++ excitebike/src/main/java/com/codenjoy/dojo/excitebike/model/elements/SpringboardElementType.java
@@ -1,10 +1,10 @@
-package com.codenjoy.dojo.excitebike.model.items;
+package com.codenjoy.dojo.excitebike.model.elements;
 
 /*-
  * #%L
  * Codenjoy - it's a dojo-like platform from developers to developers.
  * %%
- * Copyright (C) 2018 Codenjoy
+ * Copyright (C) 2018 - 2019 Codenjoy
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -22,29 +22,26 @@ package com.codenjoy.dojo.excitebike.model.items;
  * #L%
  */
 
-
 import com.codenjoy.dojo.services.printer.CharElements;
 
-public enum GameElementType implements CharElements {
+public enum SpringboardElementType implements CharElements {
 
-    //ASCII code of char in comment
-    NONE(' '),  //space
-    FENCE('■'),    //254
-
-    ACCELERATOR('>'),   //175
-    INHIBITOR('<'),    //177
-    OBSTACLE('|'),     //178
-    LINE_CHANGER_UP('▲'),  //30
-    LINE_CHANGER_DOWN('▼');  //31
+    SPRINGBOARD_LEFT_UP('╔'),
+    SPRINGBOARD_LEFT('/'),
+    SPRINGBOARD_LEFT_DOWN('╚'),
+    SPRINGBOARD_TOP('═'),
+    SPRINGBOARD_RIGHT_UP('╗'),
+    SPRINGBOARD_RIGHT('\\'),
+    SPRINGBOARD_RIGHT_DOWN('╝');
 
     final char ch;
 
-    GameElementType(char ch) {
+    SpringboardElementType(char ch) {
         this.ch = ch;
     }
 
-    public static GameElementType valueOf(char ch) {
-        for (GameElementType el : GameElementType.values()) {
+    public static SpringboardElementType valueOf(char ch) {
+        for (SpringboardElementType el : SpringboardElementType.values()) {
             if (el.ch == ch) {
                 return el;
             }
