+++ excitebike/src/main/java/com/codenjoy/dojo/excitebike/services/parse/MapParser.java
@@ -27,8 +27,8 @@ import com.codenjoy.dojo.excitebike.model.items.Fence;
 import com.codenjoy.dojo.excitebike.model.items.Inhibitor;
 import com.codenjoy.dojo.excitebike.model.items.LineChanger;
 import com.codenjoy.dojo.excitebike.model.items.Obstacle;
-import com.codenjoy.dojo.excitebike.model.items.bike.Bike;
-import com.codenjoy.dojo.excitebike.model.items.springboard.SpringboardElement;
+import com.codenjoy.dojo.excitebike.model.items.Bike;
+import com.codenjoy.dojo.excitebike.model.items.SpringboardElement;
 
 import java.util.List;
