+++ excitebike/src/test/java/com/codenjoy/dojo/excitebike/client/BoardParametrizedTest.java
@@ -22,9 +22,9 @@ package com.codenjoy.dojo.excitebike.client;
  * #L%
  */
 
-import com.codenjoy.dojo.excitebike.model.items.GameElementType;
-import com.codenjoy.dojo.excitebike.model.items.bike.BikeType;
-import com.codenjoy.dojo.excitebike.model.items.springboard.SpringboardElementType;
+import com.codenjoy.dojo.excitebike.model.elements.GameElementType;
+import com.codenjoy.dojo.excitebike.model.elements.BikeType;
+import com.codenjoy.dojo.excitebike.model.elements.SpringboardElementType;
 import com.codenjoy.dojo.services.printer.CharElements;
 import org.junit.Test;
 import org.junit.runner.RunWith;
